refactor DnsRecord to use new Resource interface
